show success banner on cookie page when cookies set and javascript disabled + tidy up
